RagPlatform/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bcrypt==4.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
docx2txt==0.9
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.135.2
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.1
greenlet==3.3.2
groq==1.1.2
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.8.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.23
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
motor==3.7.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.4
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
passlib==1.7.4
portalocker==3.2.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pymongo==4.16.0
pyparsing==3.3.2
pypdf==6.9.2
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
qdrant-client==1.17.1
regex==2026.3.32
requests==2.33.0
requests-toolbelt==1.0.0
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.4.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0